requirements.txt
Adafruit-Blinka==6.15.0
adafruit-circuitpython-neopixel==6.2.3
adafruit-circuitpython-pixelbuf==1.1.1
Adafruit-PlatformDetect==3.17.2
Adafruit-PureIO==1.1.9
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
opencv-python==4.5.4.60
pycparser==2.21
pyftdi==0.53.3
pyserial==3.5
pyusb==1.2.1
schedule==1.1.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
